Introduce logic to query account information
This patch marks the first change in the creation of apcacli, a program
for interacting with the Alpaca API. The change lays out the basic
scaffolding we are going to use for parsing command line arguments. In
particular, we use the structopt crate for declaratively describing
arguments.
In this first step we retrieve and print out basic information about the
user's account.