Add support for submitting an order
With this change we add further support for submitting an order to
Alpaca. We do so by introducing a new subcommand, order, which in turn
contains another subcommand, submit. The latter expects positional
arguments common among all orders and accepts options for configuring
in more detail what the order shall look like.