Factor out function for printing position listing
This change factors out a dedicated function for the printing of the
position listing. While at it, it also guards this function with a check
that positions are actually available, in an attempt to mimic the
behavior of the order listing (which does not display anything if there
are no orders).