Use anyhow crate for error handling
We are still not entirely happy with our error handling story. Some time
back the std::error::Error type (which ideally is implemented by each
meaningful error) added support error chaining. Ideally, we would like
to at least walk that chain for errors we received from lower level
crates. That, however, is not easily possible when we print errors at
the source, as it would make the code extremely verbose.
With this change we fall back to actually returning an error. However,
we work with an Error as provided by the anyhow crate. Those errors
preserve all chaining information and allow for easy annotation with
context.