Enable logging based tracing support
The apca crate recently got migrated to using the tracing crate for
providing tracing functionality. This change adjusts the program to
catch up to this change. In particular, instead of setting up logging
using the simplelog crate we now have to register a tracing subscriber
that logs events and spans.
One implication of the simplelog removal is that we no longer see hyper
emitted log messages. That is deemed fine (in fact, perhaps even
beneficial), as those messages were never found to be very useful.