Format prices more consistently
We have noticed one problem with the way we format "optional" prices
(such as an optional limit), in that non-present prices are formatted as
"N/A <currency>".
To fix this problem, this change adjusts the format_price function to
include the currency directly in the result. Doing so we also noticed a
couple more locations where we should have used the format_price
function but missed doing so. Lastly, having all of the price related
formatting use this function we we more easily print a thousands
separator in the future.