Colorize gains/losses
Colors can help a lot in decreasing the time it takes to understand a
listing or at least infer where to look at more closely. To that end,
this change enables the coloring of gains/losses when printing
positions.
There are various crates that allow us to colorize text in a terminal
using ANSI escape sequences, all with different trade-offs. We currently
rely on the yansi crate for doing this job. The downside is that it is a
new dependency.
Ideally we would want to use ansi_term instead (which is already pulled
in by other dependencies), but it turns out that the color enabled
strings it produces do not handle advanced formatting options as we use
them properly. See https://github.com/ogham/rust-ansi-term/issues/59 for
details.
We also have used the colored crate successfully, but it does not
support RGB colors (only predefined ones) and the default green is
commonly too bright. Also, it comes with more dependencies than
seemingly necessary.