Print total value of order in order listing
When getting a quick overview of unfilled orders it is often a
significance to understand the total value of an order. With this change
we calculate this value and include it in the order listing.
Unfortunately, due to the way this feature is implemented currently, we
do not necessarily nicely columnize this product. This is fixable, but
it will require a bit of a rework of this code.