Implement -u/--good-until argument without Option
This change reworks the way we implement the -u/--good-until argument.
So far we have used an Option and fallen back to a default if it is
None. However, the structopt crate actually supports setting a default
value and so we can directly work with a GoodUntil value instead of
having to wrap it in an Option. Doing so has at least two advantages:
first, it is more concise and second, it keeps the definition of the
parameter together with the definition of the default value to use.