Do not sort orders by symbol
So far we have explicitly sorted the list of orders that we retrieve
based on the traded symbol. At first that was found to be a reasonable
way to go about things, but a chronological sorting (which is what the
API reports back) makes much more sense long term.
Hence, this change removes this explicit sorting step from the logic.