Add support for listing account activities
It can be tremendously useful to get an overview over past trades and
other activities to understand what has happened in the past. The Alpaca
API provides a way to query "account activities" for this purpose. There
are two main types of activities: trade activities and non-trade
activities.
This change adds support for listing the past recent activities.