Print actual time for trade account activities
So far we were only printing the date on which a trade account activity
happened. But once we get more activity happening on an intra-day basis,
it becomes more and more important to understand when exactly a trade
occurred.
To that end, this change adjusts the kind of time stamp that is being
printed by also including hours, minutes, and seconds of the transaction
time.