Rename good-until to time-in-force
The term "time in force" is rather abstract and does not necessarily
convey a lot of information to users not familiar with trading on
Alpaca. For that reason we initially went with an alternative, "good
until", instead. Unfortunately, though, with the introduction of market
closing auction orders we struggled to match those order types with this
new terminology.
To that end, this change just falls back to using the term "time in
force" as it is used by Alpaca.