Fix panic when piping output to head(1)
As it turns out, the program panics when its output is piped to a
program such as head(1). The reason for this behavior is that println
panics when the underlying write(2) reports an error. This write, going
to stdout in our case, will fail with an EPIPE if the other end is
closed. The head program in turn will, presumably, close its stdin
(which is "the other end") once it has seen the desired data.
While everything is working as expected, the exhibited behavior is far
from desirable. To fix this problem, this change redefines the println
macro to safely ignore EPIPE, keeping the panic for all other errors.