Do not deserialize and serialize streamed events when emitting JSON
In the past we relied on the apca crate providing serialization support
in order to be able to emit a JSON representation of the received
account and trade events. Providing such serialization support proved to
be a burden to the point that it got removed in version 0.13. This
change updates the version we consume to 0.13 and reworks the way with
which we provide JSON output of updates.
Instead of deserializing and then serializing again, with this change we
make use of the lower-level stream function with a custom event type
that does not parse the (inner most) update event object. The object is
preserved in the form of a string and can just be printed -- without any
reserialization happening.