Use default color instead of black for indicating no gains/losses
Depending on the terminal's color theme used, black may not necessarily
be a good color to use for indicating a field that has no gains/losses
(which would be highlighted green or red, respectively), because it may
not be well visible.
With this change we merely use the default color in those cases, which
should be more appropriate.