Update minimum supported Rust version to 1.44.1
This change updates the minimum supported version of Rust to 1.44.1.
The pin-project-lite crate has included a clippy configuration that is
not understood by our current version of the program, causing the CI
pipeline to fail. Given that apca 0.17.0 already states 1.44.1 as the
minimum supported Rust version, we just tag along without bothering too
much how else we could solve that.