Print leg orders as part of 'order list' command
With this change we print any leg orders associated with an order
retrieved as part of the 'order list' command. Note that it appears that
Alpaca does not actually ever populate the leg field in order retrieved
this way. It is unclear whether that is by-design or an oversight, but
we anticipate this behavior potentially changing in the future and so we
are ready.