Use anyhow::Result alias
We have a bunch of functions that return a Result<T, Error>, where Error
is anyhow's Error type. Similar to certain modules from the standard
library, the anyhow crate already provides a Result alias that by
default uses its error type.
With this change we switch over to using the Result type from the anyhow
crate instead of explicitly providing the error type we use everywhere.