Remove account update streaming support
Streaming of account updates is not supported by Alpaca. It was
originally documented but turned out to have never been implemented in
the backend. By now, documentation surrounding the feature has been
removed.
To that end, this change removes support for streaming of account
updates via the events account subcommand from the program.