Remove --json argument from events subcommand
Version 0.22.0 of the apca crate will no longer allow us to stream "raw"
events, as we currently support for trade updates. Such support may come
back in the future, but for now we can live without it. To that end,
this change removes the --json argument from the events subcommand,
which allowed users to forgo any decoding and print the received JSON
"blobs" directly.