Overhaul and simplify CI configuration
This change overhauls our CI configuration in an attempt to simplify it
as well as speed up runs. Most importantly, we get rid of the two stages
we had so far and flatten everything. In addition to that, we split the
debug and release builds and include the test step in the debug build
job. Lastly, we remove the --verbose flag from a bunch of commands,
which was not found to be useful.