Cache artifacts in CI
Most of our CI pipeline runs are for only small and incremental changes
over a previous version of the code base. As such, caching and restoring
previous build artifacts and relying on incremental builds can yield
great speed ups over the unconditional full compilation that we
currently have.
To make this happen, this change introduces a bunch of caching to the CI
configuration. In addition to build artifacts, we also cache parts of
the cargo home directory, which would otherwise be repeatedly pulled.