Rename events subcommand to updates
This change renames the 'events' subcommand to 'updates'. In the past
the term event captured the availability of both account as well as
trade/order updates perfectly. However, in the future we want to support
streaming of realtime market data and for that the old terminology no
longer works well.