Fail gracefully when building from archive or when Git is not installed
It is actually not safe for us to unconditionally assume that we are in
a Git repository. When a user builds and installs a binary from
crates.io using cargo install, for example, the build would be triggered
on the archived source code, which does not contains version control
files. Furthermore, we probably should also not assume that the user has
Git installed and fail more gracefully if that is not the case. With
this change we take care of both cases. Note that so far we lucked out
and did not block building of the binary from crates.io because we were
missing the Git command status check that we added earlier.