Add the ability to create more bracket order types
This change adds support for specifying more bracket order types to the
program. Specifically, we recognize two new optional arguments for the
order submit sub-command: --stop-loss-stop-price and
--stop-loss-limit-price.
If only --stop-loss-stop-price is provided, we create a
one-triggers-other order with a stop loss (as opposed to a take profit,
which we already support). If both the existing --take-profit-price and
--stop-loss-stop-price are given, we will submit a "true" bracket order
with both a take profit and a stop loss leg. Lastly, if
--stop-loss-limit-price is specified additionally, the stop loss leg
will actually be a stop-limit order (as opposed to a regular stop market
order).