Stop flagging usage of deprecated functionality in clippy CI stage
With a recent release clap has deprecated a bunch of functionality that
evidently is still used by their derive logic. Because we deny warnings
in our clippy CI stage and warnings are emitted for usage of deprecated
items, we end up failing this stage.
This change allows the usage of deprecated functionality in this stage
to fix this problem.