Add proper CI job for building with minimum dependency versions
We have repeatedly been impacted by more recent versions of dependencies
requiring more recent versions of Rust in our CI workflow that tests
building with the minimum supported Rust version. That is mostly the
case because we do not lock our dependent versions via a Cargo.lock
file.
In order to prevent any issues from this setup, this change introduces a
proper CI job that makes sure to compile against the minimum versions of
dependencies with the minimum Rust version we support.