Add --begin arguments to account activity get sub-command
The account activity get sub-command currently unconditionally lists all
the account activities available. This can be somewhat unwieldy for
accounts that have been open for a while or have a lot of transactions.
To give the user a bit more control over what to display, this change
adds the --begin argument to said sub-command to specify the start date
at which to begin listing activities.