Introduce support for user-provided extensions
This change introduces support for discovering and executing
user-provided extensions to the program. Extensions are useful for
allowing users to provide additional functionality on top of apcacli
proper.
Implementation-wise we stick to an approach similar in nature to git or
cargo subcommands: we search the directories listed in the PATH
environment variable for a file that starts with "apcacli-", followed
by the extension name. This file is then executed. Note that currently
the only contract between the main program and extensions is that:
- extensions are executed when the corresponding subcommand is run
  via the main program
- the APCACLI environment variable is set that points to the apcacli
  binary being used