safeguard: Make stop price configurable
One important aspect pertaining the safety of having a stop loss is the
margin of the stop price over the limit price. This margin is typically
very dependent on the volume being traded and volatility of the asset in
question and as such, is hard to put in stone generally.
To that end, with this change we make this stop price configurable via
the new -s/--stop-price argument that expects as percentage gain value.