safeguard: Add support for providing symbols for positions to check
Sometimes we may not want to check all open positions for stop loss
orders. For example, if we have a lot of open positions that we know are
not in compliance, because we are specifically asking for an unusually
high stop price spread, that would just cause a lot of output to sift
through.
In such cases the user would know the symbols of the positions of
interest and with this change such a list is accepted as positional
arguments, causing all other positions to be skipped from the
evaluation.