safeguard: Use shorter argument value names
For whatever reason clap employs some nonsensical heuristic where it
switches to the long help text when options and their arguments exceed a
certain length (or the resulting line reaches the terminal width or
whatever). We can "fix" that with terminal width setting, except that we
are not really interested in overwriting the terminal width, we just
want to disable this stupid heuristic.
In the past we circumvented it by specifying a shorter "name" property,
but of course we forgot about these adjustments once more options were
added. As a result, the long help text was used again with the
--min-gain-percent option.
Now, it turns out that this is not exactly the property we need, because
we can't use the same name multiple times. What we really want is the
"value name".
 So with this change we set this property for the offending arguments.