Fix help text display for sub-commands
The logic we added for displaying extensions in the program's help text
turned out to be faulty when the user asked for the help to a
sub-command. Evidently, the display of such a sub-command help text is
covered by the same error kind, but it is wrong to simply display the
program help text unconditionally.
With this change we fix this state of affairs by re-triggering the
program parsing on the custom crafted command object, which will ensure
that the correct help text is printed on error.