Stop installing recommended packages in CI
We don't need "recommended" packages to be installed in CI, but that's
what apt-get chooses to do by default. Provide the
--no-install-recommends option to prevent this from happening.
While at it, also provide the --yes switch to make sure no user input is
inquired. That should be the default when no TTY is present, but it
seems best to be explicit.