Strip pre-built binaries
Our pre-built binaries are arguably excessively large, because they
carry debug and symbol information in them. That may or may not be a
reasonable default when building manually, but it may be making the
wrong trade-off for pre-built binaries.
For the time being, build those binaries with debug and symbol
information stripped.