Return ExitCode from main()
It's not great to invoke std::process::exit() directly, because a
program exited this way is not cleaned up properly: destructors may not
run and stdio buffers not flushed.
This is a potential cause of bugs and in fact it can be avoided by
returning an ExitCode from main() instead. So do just that.